% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{as_datatree}
\alias{as_datatree}
\title{As Data Tree}
\usage{
as_datatree(structure, type = c("code", "desc", "comb"), sep = ": ")
}
\arguments{
\item{structure}{A data.frame with labelled}

\item{type}{character, one of
\itemize{
\item \code{code} a data tree with levelNames based on hierarchy codes
\item \code{desc} a data tree with levelNames based on hierarchy code descriptions
\item \code{comb} a data tree with levelNames that combines the code and its description
}}

\item{sep}{character, a separate to use when combining the code and its description as the level name, default is a colon and
space \verb{: }}
}
\value{
An option of class \code{data.tree}
}
\description{
Converts a data.frame generated by a get_ function into a hierarchy using \code{\link[data.tree:as.Node]{data.tree::as.Node()}}.
}
