% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{flatten_datatree}
\alias{flatten_datatree}
\title{Flatten Data Tree}
\usage{
flatten_datatree(tree, ...)
}
\arguments{
\item{tree}{an object of class \code{data.tree}}

\item{...}{other attributes to output from the tree}
}
\value{
A data.frame with columns:
\itemize{
\item \code{from} the parent path string
\item \code{to} the child path string
\item \code{level} the tree depth of the child, where root = 1
}
}
\description{
Converts a data.tree hierarchy object into a parent, child data frame.
}
