% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{flatten_tree}
\alias{flatten_tree}
\title{Flatten Tree}
\usage{
flatten_tree(tree, ...)
}
\arguments{
\item{tree}{an object of class \code{data.tree}}

\item{...}{other attributes to output from the tree}
}
\value{
A data.frame with columns:
\itemize{
\item \code{from} the parent path string
\item \code{to} the child path string
\item \code{level} the tree depth of the child, root = 1
\item \code{parentName} the name of the parent node
\item \code{name} the name of the child node
}
}
\description{
Converts a data.tree hierarchy object into a parent, child data frame.
}
