% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_AHECC}
\alias{get_AHECC}
\title{Get AHECC}
\usage{
get_AHECC(version = c("AHECCv2017", "AHECCv2022"))
}
\arguments{
\item{version}{character string, one of:
\itemize{
\item AHECCv2017, version 2017, finalised January 2021
\item AHECCv2022, version 2022, finalised January 2022
}}
}
\value{
A data.frame with the classification hierarchy as labeled vectors
}
\description{
Downloads the Australian Harmonized Export Commodity Classifcation
}
