% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_ANZSCO}
\alias{get_ANZSCO}
\title{Get ANZSCO}
\usage{
get_ANZSCO(version = NULL)
}
\arguments{
\item{version}{character, one of
\itemize{
\item \code{ANZSCOv2021} version 2021
\item \code{ANZSCOv2022} version 2022
\item \code{OSCAv2024} version Occupation Standard Classification for Australia for 2024
}}
}
\value{
A data.frame with the classification hierarchy as labeled vectors
}
\description{
Downloads the Australian and New Zealand Standard Classification of Occupation
}
