% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_ANZSIC}
\alias{get_ANZSIC}
\title{Get ANZSIC}
\usage{
get_ANZSIC(...)
}
\value{
A data.frame with the classification hierarchy as labeled vectors
}
\description{
Downloads the Australian and New Zealand Standard Industry Classification
}
