% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_ANZSOC}
\alias{get_ANZSOC}
\title{Get ANZSOC}
\usage{
get_ANZSOC()
}
\value{
A data.frame with the classification hierarchy as labeled vectors
}
\description{
Downloads the Australian and New Zealand Standard Offence Classification
}
