% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_ANZSRC}
\alias{get_ANZSRC}
\title{Get ANZSRC}
\usage{
get_ANZSRC(version = c("ANZSRCvFoR", "ANZSRCvSEO"))
}
\arguments{
\item{version}{character, one of
\itemize{
\item \code{ANZSRCvFoR} version Field of Research
\item \code{ANZSRCvSEO} version SEO
}}
}
\value{
A data.frame with the classification hierarchy as labeled vectors
}
\description{
Download the Australian and New Zealand Standard Research Classification
}
