% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_ASCCEG}
\alias{get_ASCCEG}
\title{Get ASCCEG}
\usage{
get_ASCCEG(...)
}
\value{
A data.frame with the classification hierarchy as labeled vectors
}
\description{
Download the Australian Standard Classification of Cultural and Ethnic Groups
}
