% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_ASCDOC}
\alias{get_ASCDOC}
\title{Get ASCDOC}
\usage{
get_ASCDOC()
}
\value{
A data.frame with the classification hierarchy as labeled vectors
}
\description{
Download the Australian Standard Classification of Drugs of Concern
}
