% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_ASCED}
\alias{get_ASCED}
\title{Get ASCED}
\usage{
get_ASCED(version = c("ASCEDvLevel", "ASCEDvField"))
}
\arguments{
\item{version}{character, one of
\itemize{
\item \code{ASCEDvLevel} version Level of Education
\item \code{ASCEDvField} version Field of Education
}}
}
\value{
A data.frame with the classification hierarchy as labeled vectors
}
\description{
Download the Australian Standard Classification of Education
}
