% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_ASCL}
\alias{get_ASCL}
\title{Get ASCL}
\usage{
get_ASCL()
}
\value{
A data.frame with the classification hierarchy as labeled vectors
}
\description{
Download the Australian Standard Classification of Languages
}
