% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_ASCRG}
\alias{get_ASCRG}
\title{Get ASCRG}
\usage{
get_ASCRG(version = NULL)
}
\arguments{
\item{version}{character, one of
\itemize{
\item \code{ASCRGv2024} version 2024
\item \code{ASCRGv2016} version 2016
}}
}
\value{
A data.frame with the classification hierarchy as labeled vectors
}
\description{
Downloads the Australian Standard Classification of Religious Groups
}
