% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/un.R
\name{get_COICOP}
\alias{get_COICOP}
\title{Get COICOP}
\usage{
get_COICOP(version = c("COICOPv2018", "COICOPv1999"))
}
\arguments{
\item{version}{character, one of
\itemize{
\item \code{COICOPv2018} version 2018
\item \code{COICOPv1999} version 1999
}}
}
\value{
A data.frame with the classification hierarchy as labeled vectors
}
\description{
Downloads the United Nation's Classification of Individual Consumption According to Purpose
}
