% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_CPICC}
\alias{get_CPICC}
\title{Get CPICC}
\usage{
get_CPICC(version = NULL)
}
\arguments{
\item{version}{character, one of
\itemize{
\item \code{CPICCv16} version 16
\item \code{CPICCv15} version 15
}}
}
\value{
A data.frame with the classification hierarchy as labeled vectors
}
\description{
Download the Consumer Price Index Commodity Classification
}
