% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_OSCA}
\alias{get_OSCA}
\title{Get OSCA}
\usage{
get_OSCA(...)
}
\value{
A data.frame with the classification hierarchy as labeled vectors
}
\description{
Downloads the Occupation Standard Classification for Australia
}
