% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_SACC}
\alias{get_SACC}
\title{Get SACC}
\usage{
get_SACC()
}
\value{
A data.frame with the classification hierarchy as labeled vectors
}
\description{
Download the Standard Australian Classification of Countries
}
