% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_SESCA}
\alias{get_SESCA}
\title{Get SESCA}
\usage{
get_SESCA(version = NULL)
}
\arguments{
\item{version}{character, one of
\itemize{
\item \code{SISCAv2021} version 2021
\item \code{SISCAv2008} version 2008
}}
}
\value{
A data.frame with the classification hierarchy as labeled vectors
}
\description{
Download the Standard Economic Sector Classifications of Australia
}
