% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{unlabel}
\alias{unlabel}
\title{Unlabel}
\usage{
unlabel(structure)
}
\arguments{
\item{structure}{A data.frame generated by get_}
}
\value{
A data.frame with separate columns for hierarchy codes and code descriptions
}
\description{
Removes the labels column vectors from a hierarchy data frame
}
