% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_AHECC}
\alias{get_AHECC}
\title{Get AHECC}
\usage{
get_AHECC(version = c("AHECCv2017", "AHECCv2022"))
}
\value{
A data.frame with the classification hierarchy as labeled vectors
}
\description{
Downloads the Australian Harmonized Export Commodity Classifcation version 2017, finalised January 2021
}
