% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{unlabel}
\alias{unlabel}
\title{Unlabel}
\usage{
unlabel(structure, name = "_desc")
}
\arguments{
\item{structure}{a data.frame generated by get_}

\item{name}{string, default \verb{_desc}, a suffix to use for the added label columns}
}
\value{
A data.frame with separate columns for hierarchy codes and code descriptions
}
\description{
Removes the labels column vectors from a hierarchy data frame
}
